@@ -1,6 +1,6 @@
 {
   "name": "pollbotplus.com",
-  "version": "0.1.0",
+  "version": "0.1.0-dev.0",
   "description": "The PollBotPlus website.",
   "scripts": {
     "dev": "vite",
@@ -11,17 +11,18 @@
   "license": "",
   "dependencies": {
     "classnames": "^2.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^8.0.3",
+    "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
-    "@honkhonk/vite-plugin-svgr": "^1.1.0",
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
-    "@vitejs/plugin-react": "^1.1.1",
-    "sass": "^1.44.0",
-    "typescript": "^4.5.2",
-    "vite": "^2.7.1"
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^1.3.2",
+    "sass": "^1.54.0",
+    "typescript": "^4.7.4",
+    "vite": "^3.0.4",
+    "vite-plugin-svgr": "^2.2.1"
   }
 }
