@@ -1,65 +1,65 @@
 import "./css/index.scss";
 
-import ReactDOM from "react-dom";
-import Button from "./components/Button";
+import { useEffect } from "react";
+import { createRoot } from "react-dom/client";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Footer from "./components/Footer";
 import Header from "./components/Header";
-
-import EmbedPollImgUrl from "./img/embed-crop-tiny.png";
-import DarkEmbedPollImgUrl from "./img/dark-embed-crop-tiny.png";
+import ScrollToTop from "./components/ScrollToTop";
+import Home from "./pages/Home";
+import PrivacyPolicy from "./pages/PrivacyPolicy";
+import TermsOfService from "./pages/TermsOfService";
 
 const App = () => {
-	const INVITE_URL =
-		"https://discord.com/api/oauth2/authorize?client_id=804245390642642965&scope=applications.commands+bot&permissions=2416274496";
-	const SUPPORT_SERVER = "https://discord.gg/krZPZDD9b7";
+	useEffect(() => {
+		const browserDefaultDark = window.matchMedia(
+			"(prefers-color-scheme: dark)"
+		).matches;
+		if (
+			(browserDefaultDark && !localStorage.getItem("dark")) ||
+			(!browserDefaultDark && localStorage.getItem("dark"))
+		) {
+			// document.body.classList.add("dark");
+			document.documentElement.classList.add("dark");
+		}
+	}, []);
 
 	return (
-		<div>
-			<div className="g-wrapper">
-				<Header />
-				<div className="landing">
-					<div className="text-and-buttons">
-						<div>
-							<h2>Create quick, easy, and beautiful polls.</h2>
-							<div className="vertical-divider"></div>
-							<div className="buttons">
-								<Button
-									type="primary"
-									label="Add to Server"
-									href={INVITE_URL}
-									target="_blank"
-								/>
-								<div className="divider"></div>
-								<Button
-									type="secondary"
-									label="Support Server"
-									href={SUPPORT_SERVER}
-									target="_blank"
-								/>
-							</div>
-						</div>
-					</div>
-					<div className="image-container">
-						<img
-							src={EmbedPollImgUrl}
-							alt="A poll made by a user"
-							className="image light-image"
-							width="462"
-							height="342"
-						></img>
-						<img
-							src={DarkEmbedPollImgUrl}
-							alt="A poll made by a user"
-							className="image dark-image"
-							width="462"
-							height="342"
-						></img>
-					</div>
-				</div>
-			</div>
-			<Footer />
-		</div>
+		<BrowserRouter>
+			<ScrollToTop />
+			<Routes>
+				<Route
+					path="/"
+					element={
+						<>
+							<Home />
+							<Footer selected="home" />
+						</>
+					}
+				/>
+				<Route
+					path="/terms"
+					element={
+						<>
+							<Header />
+							<TermsOfService />
+							<Footer />
+						</>
+					}
+				/>
+				<Route
+					path="/privacy"
+					element={
+						<>
+							<Header />
+							<PrivacyPolicy />
+							<Footer />
+						</>
+					}
+				/>
+			</Routes>
+		</BrowserRouter>
 	);
 };
 
-ReactDOM.render(<App />, document.getElementById("root"));
+createRoot(document.getElementById("root")!).render(<App />);
